requirements.txt
Django==4.2.11
firebase-admin==6.5.0
djangorestframework==3.15.1
markdown==3.6
django_filter==24.2
drf-yasg
drf-yasg[validation]
python-dotenv==1.0.1
django-cors-headers
whitenoise
